// src/lib2/models/DeviseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DeviseDto = {
    id?: string;
    code: string;
    nom: string;
    symbole?: string;
    est_nationale?: boolean;
    actif?: boolean;
    created_at?: string;
};